Add multiple brush support to the paint brush tool
Pinta now supports multiple brushes that can be added through MEF. Four
brushes are currently implemented:

 - Plain (your basic line)
 - Grid
 - Circles
 - Squares

The last three are 'procedural' with more to come. Currently there is
only a basic UI for switching brushes. There is not yet a UI for
configuring brush parameters.

The brushes need to return their invalidation rectangle. Currently the
entire canvas is being invalidated. This is a regression over the
previous hard-coded brush.

Implemented QuadraticCurveTo as a Cairo.Context extension method.